+++ src/Node.js
@@ -1641,7 +1641,7 @@
      * @param {Number} [config.quality] jpeg quality.  If using an "image/jpeg" mimeType,
      *  you can specify the quality from 0 to 1, where 0 is very poor quality and 1
      *  is very high quality
-     * @paremt {Number} [config.pixelRatio] pixelRatio of ouput image url. Default is 1
+     * @param {Number} [config.pixelRatio] pixelRatio of output image url. Default is 1
      * @returns {String}
      */
     toDataURL: function(config) {